Vary relation part name based on relation class.
I.e. dependency relations have a head and dependent;
for other relations we refer to the more generic source and target.
Each relation has a relationClass (set during indexing through the
.blf.yaml config file); the relationClass is looked up in the locale
files to see if there's specific terminology for this class. The
defaults are used if not.

Other classes can be added to the locale files as needed.

Squashed commit of the following:

commit f0858636d0dcd078a738ed61ee319f98fe3b864f
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Wed Oct 30 16:31:16 2024 +0100

    relation part by class

commit 98e024f00de0ac5c83d36ed39b7497bb6354eeda
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Wed Oct 30 11:37:35 2024 +0100

    WIP